refactor: address in-scope review findings
- graduation-progress/layout.module.scss: drop background-color duplicated
  by utilities.app-layout mixin; remove now-unused color import
- errorMapping: drop raw backend error_message fallback; unknown codes now
  return DEFAULT_ERROR_MESSAGE so internal texts never surface to users
- idempotency.ts: move out of src/shared/api/ (api:gen boundary) to
  src/shared/utils/; update imports in PortalLoginForm and resync/login
- scraping/page.tsx: stop sending raw studentCode (학번) to Sentry.setUser;
  use server-issued jobId as opaque user id for issue grouping
