+++ src/app/resync/login/page.tsx
@@ -8,7 +8,7 @@ import { ROUTES } from '@/constants/routes';
 import { useInternalRouter } from '@/hooks/useInternalRouter';
 import { usePortalLinkMutation } from '@/features/portal-link/hooks';
 import { RESYNC_JOB_ID_KEY } from '@/constants/portal-link';
-import { generateIdempotencyKey } from '@/shared/api/utils/idempotency';
+import { generateIdempotencyKey } from '@/shared/utils/idempotency';
 import { FunnelHeadline, SchoolCard } from '../../(funnel)/components';
 import styles from './page.module.scss';
